offline-transcriber/requirements.txt
numpy>=1.20.0
pydub>=0.25.1
openai-whisper>=20230314
tqdm>=4.64.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.10.0
flake8>=6.0.0 